requirements.txt
pillow
requests
lxml
flask
numpy
Werkzeug==2.2.2
PyQt6
torch
pyperclip
